Refactor Activity class and Parser class to packages
