+++ src/main/java/seedu/busybreak/activity/BudgetPlan.java
@@ -1,5 +1,4 @@
 package seedu.busybreak.activity;
-import seedu.busybreak.Activity;
 
 import java.util.ArrayList;
 import java.util.Locale;
