+++ src/main/java/seedu/busybreak/parser/Parser.java
@@ -1,10 +1,12 @@
-package seedu.busybreak;
+package seedu.busybreak.parser;
+
+import seedu.busybreak.Ui;
 
 import java.util.Arrays;
 import java.util.Scanner;
 
 public class Parser {
-    static GetCommand parseUserInput() {
+    public static GetCommand parseUserInput() {
         Scanner in = getScanner();
         if (in == null) {
             return null;
@@ -29,11 +31,11 @@ public class Parser {
         return in;
     }
 
-    public static int parseActivityIndex(String activityIndexString){
+    public static int parseActivityIndex(String activityIndexString) {
         return Integer.parseInt(activityIndexString) - 1;
     }
 
-    public static ParseEditDetails parseEditActivityDetails(String[] userInputArray){
+    public static ParseEditDetails parseEditActivityDetails(String[] userInputArray) {
         String[] inputDetailsArray = Arrays.copyOfRange(userInputArray, 2, userInputArray.length);
         String inputDetails = String.join(" ", inputDetailsArray).trim();
         String[] editDetails = inputDetails.split("\\s+(?=desc/|d/|t/|c/)");
@@ -71,7 +73,8 @@ public class Parser {
     }
 
     public record ParseEditDetails(String date, String time, String description, String cost,
-                                   boolean hasInvalidDetail){}
+                                   boolean hasInvalidDetail) {
+    }
 
     public static ParseActivityData getParseActivityData(String[] userInputArray) {
 
@@ -149,6 +152,6 @@ public class Parser {
     public record ParseActivityData(String date, String time, String description, String cost) {
     }
 
-    record GetCommand(String userInput, String[] userInputArray, String command) {
+    public record GetCommand(String userInput, String[] userInputArray, String command) {
     }
 }
